Update
Improve scene dynamic loading.
Static lighting fixed.
Fix resize render targets.
Major improvement for Settings service.
Improvement for the Arguments service.
Remove singletons usage.
Resource manager Improvement.
Cross-platform fixes.
Major CMakeList.txt update for MSVC.
Remove sub-module 'dependencies/googletest' for FetchContent logic.
Unify compilation options.
New method to identify observables.

+++ src/Libs/VertexFactory/TriangleGenerator.hpp
@@ -246,7 +246,7 @@ namespace EmEn::Libs::VertexFactory
 				}
 
 				/* Generate triangles. */
-				Shape< number_t > shape{triangles.size() * 3, 1, triangles.size()};
+				Shape< number_t > shape{static_cast< uint32_t >(triangles.size()) * 3U, 1U, static_cast< uint32_t >(triangles.size())};
 
 				ShapeBuilder< number_t > builder{shape};
